Corrected the import namespace and got all the spring boot components scanned

+++ src/main/java/so/fine/codingchallenge/endpoints/LeadEndpoint.java
@@ -18,7 +18,7 @@ public class LeadEndpoint {
     }
 
     @GetMapping
-    public Lead getLead(@RequestParam Integer id){
+    public Lead getLead(@RequestParam Long id){
         return leadService.getLead(id);
     }
